+++ classes/class-wc-emspay-afterpay.php
@@ -79,8 +79,8 @@ class WC_Emspay_AfterPay extends WC_Emspay_Gateway
     {
         WC()->cart->empty_cart();
 
-        echo "<p><b>".__('Your payment using AfterPay is successful.', WC_Emspay_Helper::DOMAIN)."</b></p>";
-        echo "<p>".__('The invoice will be sent to your email.', WC_Emspay_Helper::DOMAIN)."</p>";
+        echo "<p><b>".esc_html__('Your payment using AfterPay is successful.', WC_Emspay_Helper::DOMAIN)."</b></p>";
+        echo "<p>".esc_html__('The invoice will be sent to your email.', WC_Emspay_Helper::DOMAIN)."</p>";
     }
 
     /**
@@ -92,7 +92,7 @@ class WC_Emspay_AfterPay extends WC_Emspay_Gateway
             return null;
         }
 
-        echo '<fieldset><legend>'.__('Additional Information', WC_Emspay_Helper::DOMAIN).'</legend >';
+        echo '<fieldset><legend>'.esc_html__('Additional Information', WC_Emspay_Helper::DOMAIN).'</legend >';
 
         woocommerce_form_field('gender', array(
             'type' => 'select',